test/requirements.txt
pytest
requests
kafka-python
pyyaml
jinja2
jsonpath